Move non-view JxXxx classes into Js namespace
