+++ src/Js/JsExpressionable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Atk4\Ui;
+namespace Atk4\Ui\Js;
 
 /**
  * Implements a class that can be mapped into arbitrary JavaScript expression.
